+++ test/controllers/categorisation/CyaSupplementaryUnitControllerSpec.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.categorisation
 
 import base.SpecBase
 import base.TestConstants.{testEori, testRecordId, userAnswersId}
 import connectors.GoodsRecordConnector
-import controllers.categorisation.CyaSupplementaryUnitController
+import controllers.routes
 import models.helper.SupplementaryUnitUpdateJourney
 import models.{NormalMode, SupplementaryRequest, UserAnswers}
 import org.apache.pekko.Done
